Refactor all files into separate folders/packages.
We leave the docs_preprocessor and cli files as top-level members of
the main package since they reference all the other packages.
