+++ docs-preprocessor/docs_generator/generator.go
@@ -1,0 +1,3 @@
+package generator
+
+func GenerateDocs(opts *Opts)
