+++ docs-preprocessor/docs_preprocessor.go
@@ -4,18 +4,17 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	//"regexp"
-	//"strings"
-	//"io/ioutil"
-	//"errors"
 	"github.com/gruntwork-io/docs/docs-preprocessor/docfile"
 	"github.com/gruntwork-io/docs/docs-preprocessor/file"
 	"github.com/gruntwork-io/docs/docs-preprocessor/errors"
 	"github.com/gruntwork-io/docs/docs-preprocessor/logger"
+	"github.com/gruntwork-io/docs/docs-preprocessor/globs"
 )
 
+
+
 // This function will walk all the files specified in opt.InputPath and relocate them to their desired folder location
-func ProcessDocs(opts *Opts) error {
+func PreprocessDocs(opts *Opts) error {
 	return filepath.Walk(opts.InputPath, func(path string, info os.FileInfo, err error) error {
 		relPath, err := file.GetPathRelativeTo(path, opts.InputPath)
 		if err != nil {
@@ -44,7 +43,7 @@ func ProcessDocs(opts *Opts) error {
 
 // Return true if this is a file or folder we should skip completely in the processing step.
 func shouldSkipPath(path string, opts *Opts) bool {
-	return path == opts.InputPath || MatchesGlobs(path, opts.Excludes)
+	return path == opts.InputPath || globs.MatchesGlobs(path, opts.Excludes)
 }
 
 //
