+++ docs-preprocessor/globs/globs_test.go
@@ -1,4 +1,4 @@
-package main
+package globs
 
 import (
 	"testing"
