+++ docs-preprocessor/util/util.go
@@ -1,7 +1,7 @@
-package main
+package util
 
 // Return true if `slice` contains the given `strToFind`; missing generics about now...
-func strSliceContains(slice []string, strToFind string) bool {
+func StrSliceContains(slice []string, strToFind string) bool {
 	for _, str := range slice {
 		if str == strToFind {
 			return true
